setup initial typescript + eslint config
