+++ .eslintrc
@@ -1,5 +1,17 @@
 {
-    "parser": "babel-eslint",
+    "parser": "@typescript-eslint/parser",
+    "extends": [
+        "eslint:recommended",
+        "plugin:@typescript-eslint/recommended",
+        "plugin:react/recommended",
+        "plugin:import/errors",
+        "plugin:import/warnings"
+    ],
+    "parserOptions": {
+        "ecmaVersion": 2020,
+        "sourceType": "module" // Allows for the use of imports
+      },
+    "plugins": ["@typescript-eslint"],
     "rules": {
         "strict": 0,
         "camelcase": 0,
@@ -63,12 +75,6 @@
             }
         ]
     },
-    "extends": [
-        "eslint:recommended",
-        "plugin:react/recommended",
-        "plugin:import/errors",
-        "plugin:import/warnings"
-    ],
     "env": {
         "es6": true,
         "browser": true,
