+++ gatsby-config.js
@@ -170,7 +170,7 @@ module.exports = {
             resolve: 'gatsby-plugin-eslint',
             options: {
                 stages: ['develop'],
-                extensions: ['js'],
+                extensions: ['js', 'ts', 'jsx', 'tsx'],
                 exclude: ['node_modules', '.cache', 'public'],
             },
         },
