+++ package.json
@@ -46,13 +46,16 @@
         "webpack": "^5.39.0"
     },
     "devDependencies": {
+        "@types/styled-components": "^5.1.11",
+        "@typescript-eslint/eslint-plugin": "^4.29.0",
+        "@typescript-eslint/parser": "^4.29.0",
         "babel-eslint": "10.1.0",
         "babel-plugin-styled-components": "^1.12.0",
         "eslint": "^7.28.0",
         "eslint-loader": "^4.0.2",
         "eslint-plugin-import": "^2.23.4",
         "eslint-plugin-react": "^7.24.0",
-        "eslint-webpack-plugin": "^2.5.4",
+        "eslint-webpack-plugin": "^3.0.1",
         "gatsby-plugin-eslint": "^3.0.0",
         "gatsby-plugin-styled-components": "^4.7.1",
         "gh-pages": "^3.2.1",
@@ -67,7 +70,8 @@
         "stylelint-config-standard": "^22.0.0",
         "stylelint-config-styled-components": "^0.1.1",
         "stylelint-processor-styled-components": "^1.10.0",
-        "stylelint-webpack-plugin": "^2.2.0"
+        "stylelint-webpack-plugin": "^2.2.0",
+        "typescript": "^4.3.5"
     },
     "keywords": [
         "deriv",
